pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openetruscan"
version = "0.4.0"
description = "Open-source tools for ancient epigraphy — built for Etruscan, designed to be copied."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "OpenEtruscan Contributors"},
]
keywords = [
    "etruscan",
    "epigraphy",
    "ancient-languages",
    "digital-humanities",
    "normalization",
    "old-italic",
    "tei",
    "epidoc",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing :: Linguistic",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "psycopg2-binary>=2.9",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "defusedxml>=0.7.1",
]

[project.optional-dependencies]
epidoc = ["lxml>=4.9"]
corpus = ["openetruscan[epidoc]"]
prosopography = ["networkx>=3.0"]
server = ["fastapi", "uvicorn", "pydantic", "pydantic-settings", "slowapi>=0.1.9"]
stats = ["numpy>=1.24", "scipy>=1.11", "scikit-learn>=1.3", "joblib>=1.3"]
lod = ["httpx>=0.25"]
neural = ["torch>=2.0", "onnxscript>=0.2"]
neural-inference = ["onnxruntime>=1.16"]
sota = ["transformers>=4.30.0", "accelerate>=0.21.0"]
all = ["openetruscan[epidoc,prosopography,server,stats,lod,neural,sota]"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "aiosqlite>=0.19",
    "httpx>=0.25",
    "ruff>=0.4",
    "semgrep>=1.60",
    "safety>=3.0",
    "openetruscan[all]",
]

[project.scripts]
openetruscan = "openetruscan.core.cli:main"

[project.urls]
Homepage = "https://openetruscan.com"
Documentation = "https://github.com/Eddy1919/openEtruscan#readme"
Repository = "https://github.com/Eddy1919/openEtruscan"
Issues = "https://github.com/Eddy1919/openEtruscan/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/openetruscan"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.ruff]
target-version = "py310"
line-length = 100
exclude = ["scripts/"]

[tool.ruff.lint]
select = ["F", "E", "B", "UP"]
ignore = ["B904", "E501", "F821", "B008"]

[tool.mypy]
python_version = "3.10"
strict = false
ignore_errors = false
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true
show_column_numbers = true
exclude = [
    "tests/",
    ".venv/",
    "venv/",
    ".git/",
    "__pycache__/",
    "build/",
    "dist/",
]

[[tool.mypy.overrides]]
module = [
    "lxml.*",
    "psycopg2.*",
    "networkx.*",
    "fastapi.*",
    "uvicorn.*",
    "pydantic.*",
    "slowapi.*",
    "numpy.*",
    "scipy.*",
    "sklearn.*",
    "joblib.*",
    "httpx.*",
    "torch.*",
    "onnxscript.*",
    "onnxruntime.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "openetruscan.api.server",
    "openetruscan.ml.classifier",
    "openetruscan.core.prosopography",
    "openetruscan.core.artifacts",
    "openetruscan.ml.neural",
]
ignore_errors = true

[tool.pyright]
typeCheckingMode = "basic"
reportArgumentType = "error"
reportGeneralTypeIssues = "error"
reportOperatorIssue = "error"
reportOptionalMemberAccess = "error"
